dieu chinh

+++ app/Http/Controllers/UserController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Http\Request;
 use App\Models\User;
@@ -19,7 +20,7 @@ class UserController extends Controller
     // }
     public function __construct(UserRepositoryInterface $userRepository)
     {
-        $this->userRepository =$userRepository;
+        $this->userRepository = $userRepository;
     }
     public function showUserManager()
     {
@@ -49,8 +50,8 @@ class UserController extends Controller
             'password' => ['required', 'string', 'min:6']
         ], [
             'name.regex' => 'Tên không được chứa ký tự đặc biệt.',
-            'name.required'=>'Tên không được để trống',
-            'email.required'=>'email không được để trống',
+            'name.required' => 'Tên không được để trống',
+            'email.required' => 'email không được để trống',
             'email.regex' => 'Email không được chứa ký tự đặc biệt.',
             'password.required' => 'Mật khẩu là bắt buộc.',
             'password.min' => 'Mật khẩu phải có ít nhất 6 ký tự.'
@@ -68,17 +69,17 @@ class UserController extends Controller
                 'message' => 'Email đã tồn tại. Vui lòng nhập email khác.'
             ], 409, [], JSON_UNESCAPED_UNICODE);
         }
-        $data=$validator->validated();
-        $createUser =[
-            'name'=>$data['name'],
-            'email'=>$data['email'],
-            'group_role'=>$data['group_role'],
-            'is_active'=>$data['is_active'],
-            'password'=>Hash::make($data['password']),
-            'is_delete'=>0,
+        $data = $validator->validated();
+        $createUser = [
+            'name' => $data['name'],
+            'email' => $data['email'],
+            'group_role' => $data['group_role'],
+            'is_active' => $data['is_active'],
+            'password' => Hash::make($data['password']),
+            'is_delete' => 0,
         ];
         $user = $this->userRepository->create($createUser);
-        
+
         return response()->json([
             'status' => 'success',
             'message' => 'Người dùng đã được thêm thành công.',
@@ -92,13 +93,13 @@ class UserController extends Controller
             'email' => ['required', 'regex:/^[a-zA-Z0-9\s@._\-]*$/u'],
             'group_role' => ['required', 'string'],
             'is_active' => ['required', 'boolean'],
-            'password'=>['nullable','string', 'min:6']
+            'password' => ['nullable', 'string', 'min:6']
         ], [
             'name.regex' => 'Tên không được chứa ký tự đặc biệt.',
             'email.regex' => 'Email không được chứa ký tự đặc biệt.',
-            'name.required'=>'Tên không được để trống',
-            'email.required'=>'email không được để trống',
-            'password.min'=>'Mật khẩu tối thiểu 6 kí tự'
+            'name.required' => 'Tên không được để trống',
+            'email.required' => 'email không được để trống',
+            'password.min' => 'Mật khẩu tối thiểu 6 kí tự'
         ]);
         $currentUser = $this->userRepository->find($id);
         if ($validator->stopOnFirstFailure()->fails()) {
@@ -124,16 +125,16 @@ class UserController extends Controller
                 'message' => 'Người dùng không tồn tại.'
             ], 404, [], JSON_UNESCAPED_UNICODE);
         }
-        $updateData =[
-            'name'=>$data['name'],
-            'email'=>$data['email'],
-            'group_role'=>$data['group_role'],
-            'is_active'=>$data['is_active']
+        $updateData = [
+            'name' => $data['name'],
+            'email' => $data['email'],
+            'group_role' => $data['group_role'],
+            'is_active' => $data['is_active']
         ];
         if (!empty($data['password'])) {
             $updateData['password'] = Hash::make($data['password']);
         }
-        $updateUser = $this->userRepository->update($id,$updateData);
+        $updateUser = $this->userRepository->update($id, $updateData);
         return response()->json([
             'status' => 'success',
             'message' => 'Người dùng đã được cập nhật thành công.',
@@ -142,6 +143,22 @@ class UserController extends Controller
     }
     public function getUserById($id)
     {
+        $data = ['id' => $id];
+        $validator = Validator::make($data, [
+            'id' => ['required', 'integer'],
+        ], [
+            'id.required' => 'Id không được để trống',
+            'id.integer' => 'Id phải là dạng số',
+        ]);
+
+        // Kiểm tra validator
+        if ($validator->fails()) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Dữ liệu không hợp lệ',
+                'errors' => $validator->errors()->toArray(),
+            ], 422, [], JSON_UNESCAPED_UNICODE);
+        }
         $user = $this->userRepository->find($id);
 
         if (!$user) {
@@ -165,7 +182,8 @@ class UserController extends Controller
             'is_active' => ['nullable', 'boolean']
         ], [
             'name.regex' => 'Tên không được chứa ký tự đặc biệt.',
-            'email.regex' => 'Email không được chứa ký tự đặc biệt.'
+            'email.regex' => 'Email không được chứa ký tự đặc biệt.',
+            'is_active.boolean' => 'Trạng thái không đúng định dạng'
         ]);
 
         if ($validator->stopOnFirstFailure()->fails()) {
@@ -191,6 +209,22 @@ class UserController extends Controller
     }
     public function deleteUser($id)
     {
+        $data = ['id' => $id];
+        $validator = Validator::make($data, [
+            'id' => ['required', 'integer'],
+        ], [
+            'id.required' => 'Id không được để trống',
+            'id.integer' => 'Id phải là dạng số',
+        ]);
+
+        // Kiểm tra validator
+        if ($validator->fails()) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Dữ liệu không hợp lệ',
+                'errors' => $validator->errors()->toArray(),
+            ], 422, [], JSON_UNESCAPED_UNICODE);
+        }
         try {
             $this->userRepository->delete($id);
             return response()->json([
@@ -231,6 +265,22 @@ class UserController extends Controller
     }
     public function blockUser($id)
     {
+        $data = ['id' => $id];
+        $validator = Validator::make($data, [
+            'id' => ['required', 'integer'], 
+        ], [
+            'id.required' => 'Id không được để trống',
+            'id.integer' => 'Id phải là dạng số',
+        ]);
+
+        // Kiểm tra validator
+        if ($validator->fails()) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Dữ liệu không hợp lệ',
+                'errors' => $validator->errors()->toArray(),
+            ], 422, [], JSON_UNESCAPED_UNICODE);
+        }
         $user = User::where('is_delete', 0)->find($id);
 
         if (!$user) {
